re PR tree-optimization/50717 (Silent code gen fault with incorrect widening of multiply)
2011-10-18  Andrew Stubbs  <ams@codesourcery.com>

	PR tree-optimization/50717

	gcc/
	* tree-ssa-math-opts.c (is_widening_mult_p): Remove the 'type'
	parameter.  Calculate 'type' from stmt.
	(convert_mult_to_widen): Update call the is_widening_mult_p.
	(convert_plusminus_to_widen): Likewise.

	gcc/testsuite/
	* gcc.dg/pr50717-1.c: New file.
	* gcc.target/arm/wmul-12.c: Correct types.
	* gcc.target/arm/wmul-8.c: Correct types.

From-SVN: r180164

+++ gcc/testsuite/gcc.target/arm/wmul-8.c
@@ -5,7 +5,7 @@
 long long
 foo (long long a, int *b, int *c)
 {
-  return a + *b * *c;
+  return a + (long long)*b * *c;
 }
 
 /* { dg-final { scan-assembler "smlal" } } */
